frontend/api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.9
mangum==0.17.0
aiofiles==23.2.1
numpy
Pillow
